Rename statechart to service and move service.js into service folder
