+++ visualizer/src/service/canvasMachine.js
