+++ visualizer/src/service/imageMachine.js
