+++ visualizer/src/service/raw/channelMachine.js
