+++ visualizer/src/service/raw/colorMachine.js
