+++ visualizer/src/service/raw/grayscaleMachine.js
