+++ visualizer/src/service/tools/autofitMachine.js
