+++ visualizer/src/service/tools/brushMachine.js
