+++ visualizer/src/service/tools/deleteMachine.js
