+++ visualizer/src/service/tools/thresholdMachine.js
