+++ visualizer/src/service/tools/toolUtils.js
