+++ visualizer/src/service/tools/trimMachine.js
